+++ ChangeLog.md
@@ -1,5 +1,12 @@
 ## change log ##
 
+* 2016-03-06: 0.3.1:
+  - skip testing of TripleStoreServiceAgRest, SelfHostedOntController
+  - more org tests
+  - some refactoring to facilitate triple store route testing
+  - add basic test for "self-hosted" request
+  - add tests for requested formats
+  
 * 2016-03-03: 0.3.1:
   - add scoverage
   - default warn log level